Refactor track blocks
