+++ app/assets/javascripts/3dbio_viewer/src/webapp/components/protvista/JumpToButton.tsx
@@ -2,17 +2,13 @@ import React from "react";
 import _ from "lodash";
 import i18n from "../../utils/i18n";
 import { Dropdown, DropdownProps } from "../dropdown/Dropdown";
+import { blockDefs } from "./protvista-blocks";
 
 export interface JumpToButtonProps {}
 
 export const JumpToButton: React.FC<JumpToButtonProps> = () => {
     const items: DropdownProps["items"] = React.useMemo(() => {
-        return [
-            { id: "generalInfo", text: i18n.t("General information") },
-            { id: "structuralInfo", text: i18n.t("Structural information") },
-            { id: "mapValidation", text: i18n.t("Map validation") },
-            { id: "ppiViewer", text: i18n.t("PPI viewer") },
-        ];
+        return blockDefs.map(block => ({ id: block.id, text: block.title }));
     }, []);
 
     return <Dropdown text={i18n.t("Jump to")} items={items} onClick={goToElement} />;
